refactor to support multiple implementors
